split JMC Agents endpoints by templates vs probes
